Replace ModelSummarizer with torchinfo and profiler
